feat(blueprint-build-utils): utilities lib for build scripts
