+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
+    "blueprint-build-utils",
     "blueprint-metadata",
     "blueprints/incredible-squaring",
     "blueprints/incredible-squaring-eigenlayer",
@@ -54,6 +55,7 @@ tangle-raw-event-listener-blueprint = { path = "./blueprints/tangle-raw-event-li
 gadget-blueprint-proc-macro = { path = "./macros/blueprint-proc-macro", default-features = false, version = "0.2.3" }
 gadget-blueprint-proc-macro-core = { path = "./macros/blueprint-proc-macro-core", default-features = false, version = "0.1.5" }
 gadget-context-derive = { path = "./macros/context-derive", default-features = false, version = "0.1.3" }
+blueprint-build-utils = { path = "./blueprint-build-utils", default-features = false, version = "0.1.0" }
 blueprint-metadata = { path = "./blueprint-metadata", default-features = false, version = "0.1.6" }
 cargo-tangle = { path = "./cli", version = "0.2.1" }
 cargo_metadata = { version = "0.18.1" }
