+++ blueprint-build-utils/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "blueprint-build-utils"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+
+[lints]
+workspace = true
