+++ blueprints/incredible-squaring-eigenlayer/Cargo.toml
@@ -59,6 +59,9 @@ num-bigint = { workspace = true }
 blueprint-test-utils = { workspace = true }
 gadget-io = { workspace = true }
 
+[build-dependencies]
+blueprint-build-utils = { workspace = true }
+
 [features]
 default = ["std"]
 std = []
